refactor coffee store page
