+++ pages/_app.js
@@ -1,38 +1,5 @@
-import { createContext, useReducer } from "react";
 import "../styles/globals.css";
-
-export const StoreContext = createContext();
-
-// TODO : use-Reducer start -------------------------------------------
-
-export const ACTION_TYPES = {
-  SET_LAT_LONG: "SET_LAT_LONG",
-  SET_COFFEE_STORES: "SET_COFFEE_STORES",
-}
-
-const storeReducer = (state, action) => {
-  switch (action.type) {
-    case ACTION_TYPES.SET_LAT_LONG: {
-      return { ...state, latLong: action.payload.latLong };
-    }
-    case ACTION_TYPES.SET_COFFEE_STORES: {
-      return { ...state, coffeeStores: action.payload.coffeeStores };
-    }
-    default: throw new Error(`Un-handle action type ${action.type}`);
-  }
-};
-
-const StoreProvider = ({ children }) => {
-  const initialState = {
-    latLong: "",
-    coffeeStore: [],
-  };
-
-  const [state, dispatch] = useReducer(storeReducer, initialState);
-  return (
-  <StoreContext.Provider value={{ state, dispatch }}>{children}</StoreContext.Provider>
-  );
-}
+import StoreProvider from '../store/store-context';
 
 function MyApp({ Component, pageProps }) {
   return (
