+++ utils/index.js
@@ -1,0 +1,3 @@
+export const isEmpty = (obj) => {
+    return Object.keys(obj).length===0;
+}
